resolve cycle; move Instrument and Measurement up a package
