--- metric2/instrument/instrument.go
@@ -1,4 +1,0 @@
-package instrument
-
-type Instrument interface {
-}
