+++ metric2/sync/int64/syncint64metric.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/metric2/batch"
+	metric "go.opentelemetry.io/otel/metric2"
 )
 
 // TODO instrument options
@@ -42,14 +42,14 @@ func (u UpDownCounter) Add(ctx context.Context, x int64, attrs ...attribute.KeyV
 func (h Histogram) Record(ctx context.Context, x int64, attrs ...attribute.KeyValue) {
 }
 
-func (c Counter) Measure(x int64) batch.Measurement {
-	return batch.Measurement{}
+func (c Counter) Measure(x int64) metric.Measurement {
+	return metric.Measurement{}
 }
 
-func (u UpDownCounter) Measure(x int64) batch.Measurement {
-	return batch.Measurement{}
+func (u UpDownCounter) Measure(x int64) metric.Measurement {
+	return metric.Measurement{}
 }
 
-func (h Histogram) Measure(x int64) batch.Measurement {
-	return batch.Measurement{}
+func (h Histogram) Measure(x int64) metric.Measurement {
+	return metric.Measurement{}
 }
